agents/requirements.txt
azure-functions==1.21.3
azure-functions-durable==1.2.9
azure-identity==1.19.0
azure-keyvault-secrets==4.9.0
azure-mgmt-datafactory==9.0.0
azure-mgmt-synapse==2.0.0
azure-storage-blob==12.19.0
azure-storage-file-datalake==12.14.0
openai==1.58.1
pyodbc==5.2.0
pydantic==2.10.4
requests==2.32.3
openpyxl==3.1.5
botbuilder-core==4.16.2
botbuilder-schema==4.16.2
botbuilder-integration-aiohttp==4.16.2
azure-search-documents==11.6.0
numpy==1.26.4
